fix errors.

+++ handler/config.go
@@ -1,11 +1,11 @@
 package handler
 
 type Limitation struct {
-	MaxLimit            int `bson:"max_limit"              json:"max_limit"`
-	MaxEventTags        int `bson:"max_event_tags"         json:"max_event_tags"`         // todo(@zig)::: move to server.
-	MaxContentLength    int `bson:"max_content_length"     json:"max_content_length"`     // todo(@zig)::: move to server.
-	CreatedAtLowerLimit int `bson:"created_at_lower_limit" json:"created_at_lower_limit"` // todo(@zig)::: move to server.
-	CreatedAtUpperLimit int `bson:"created_at_upper_limit" json:"created_at_upper_limit"` // todo(@zig)::: move to server.
+	MaxLimit            uint16 `bson:"max_limit"              json:"max_limit"`
+	MaxEventTags        uint   `bson:"max_event_tags"         json:"max_event_tags"`         // todo(@zig)::: move to server.
+	MaxContentLength    uint   `bson:"max_content_length"     json:"max_content_length"`     // todo(@zig)::: move to server.
+	CreatedAtLowerLimit uint   `bson:"created_at_lower_limit" json:"created_at_lower_limit"` // todo(@zig)::: move to server.
+	CreatedAtUpperLimit uint   `bson:"created_at_upper_limit" json:"created_at_upper_limit"` // todo(@zig)::: move to server.
 }
 
 type Config struct {
